requirements.txt
matplotlib~=3.3.3
tqdm~=4.46.0
tensorflow~=2.1.0
flask~=1.1.2
numpy~=1.19.3
pandas~=1.0.3
pillow~=7.1.2
opencv-python~=4.2.0.34
gunicorn